+++ RELEASE_NOTES.md
@@ -1,5 +1,9 @@
 # Release Notes
 
+## Version 1.4.12 - Community 1.0.0
+    * Clients & Groups
+        * [HQX-120] - Display client status transitions
+
 ## Version 1.4.11 - Community 1.0.0
     * Savings Accounts & Products
         * [FOX-212] - Add inScope field to LoanProduct entity to allow isolation of loan products in scope for field operations for summary calculations